pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sap-mcp-server"
version = "1.0.0"
description = "MCP Server for SAP HANA and SQL databases"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "CData Software" }
]
keywords = ["mcp", "sap", "hana", "ai", "llm", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
hana = ["hdbcli>=2.0.0"]
odbc = ["pyodbc>=4.0.0"]
all = ["hdbcli>=2.0.0", "pyodbc>=4.0.0"]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0"]

[project.scripts]
sap-mcp = "sap_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sap_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
